requirements.txt
streamlit
pinecone
llama-index
openai
anthropic
python-dotenv
llama-index-vector-stores-pinecone
llama-index-llms-anthropic
llama-index-embeddings-openai